refactor NewPerpetualForcedActionPage
